fix(backend,frontend): fix broken Post SEO Settings, add AI SEO generation, fix /api/health
SEO Settings on the Post admin form never worked: CreatePost/EditPost only
handled metadata.* form data, never seo.* — so SEO Title/Focus Keyphrase/
Meta Description/Social fields were silently discarded on save (never
written to the seo_metadata table via Post's HasSeo relation), Edit never
re-filled them on reopen, Api\PostResource never exposed them, and the
frontend blog generateMetadata() never read them (used raw title/content
slice, no Open Graph tags at all). Fixed all four layers, verified the
save/read round-trip via tinker before wiring the frontend.

Also added a "Generate with AI" button on the SEO Settings section —
AiSeoGeneratorService calls Claude Opus 5 via the official Anthropic PHP
SDK with a JSON-schema structured output (seo_title/focus_keyphrase/
meta_description/social_title/social_description) from the post title +
content, same DB-Setting-overrides-.env credential pattern as every other
integration in this codebase. Credentials optional — the button surfaces a
clear error if ANTHROPIC_API_KEY / Setting is not configured, no crash.

Unrelated one-line fix bundled in: /api/health was permanently reporting
"degraded" due to a strict type comparison against a string Redis returns
via predis (Cache::get() !== 1 is always true since predis returns "1").

+++ backend/routes/api.php
@@ -36,7 +36,7 @@ Route::get('/health', function () {
     // Cache
     try {
         Cache::put('_health_check', 1, 5);
-        if (Cache::get('_health_check') !== 1) {
+        if ((int) Cache::get('_health_check') !== 1) {
             $allOk = false;
         }
     } catch (Exception $e) {
